fix(cli): separate option modules
